+++ test_agent/CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.1.0
+
+- Udpated unit test and widget test template prompt
+- Included more references that can be added